= 6.3.7 = * Improvement: All URLs are now checked against the Wordfence Domain Blacklist in addition to Google's. * Improvement: Better page load performance for multisite installations with thousands of tables. * Improvement: Updated the bundled GeoIP database. * Improvement: Integrated blacklist blocking statistics into the dashboard for Premium users. * Fix: Added locking to the automatic update process to ensure non-standard crons don't break Wordfence. * Fix: Fixed an activation error on multisite installations on very old WordPress versions. * Fix: Adjusted the behavior of the blacklist toggle for Free users.
git-svn-id: https://plugins.svn.wordpress.org/wordfence/trunk@1645135 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ lib/wordfenceClass.php
@@ -92,7 +92,7 @@ class wordfence {
 			$schema->dropAll();
 			wfConfig::updateTableExists();
 			foreach(array('wordfence_version', 'wordfenceActivated') as $opt){
-				if (is_multisite()) {
+				if (is_multisite() && function_exists('delete_network_option')) {
 					delete_network_option(null, $opt);
 				}
 				delete_option($opt);
@@ -167,7 +167,8 @@ class wordfence {
 	public static function dailyCron(){
 		$api = new wfAPI(wfConfig::get('apiKey'), wfUtils::getWPVersion());
 		try {
-			$keyData = $api->call('ping_api_key');
+			$postParams = wfDashboard::updatePOSTParams();
+			$keyData = $api->call('ping_api_key', array(), $postParams);
 			if(isset($keyData['_isPaidKey']) && $keyData['_isPaidKey']){
 				$keyExpDays = $keyData['_keyExpDays'];
 				$keyIsExpired = $keyData['_expired'];
@@ -208,7 +209,7 @@ class wordfence {
 			}
 			if (isset($keyData['dashboard'])) {
 				wfConfig::set('lastDashboardCheck', time());
-				wfDashboard::processDashboardResponse($keyData['dashboard']);
+				wfDashboard::processDashboardResponse($keyData['dashboard'], @base64_decode($postParams['topBlacklist']));
 			}
 		}
 		catch(Exception $e){
@@ -364,8 +365,8 @@ class wordfence {
 		if (function_exists('ignore_user_abort')) {
 			ignore_user_abort(true);
 		}
-		$previous_version = (is_multisite() ? get_network_option(null, 'wordfence_version', '0.0.0') : get_option('wordfence_version', '0.0.0'));
-		if (is_multisite()) {
+		$previous_version = ((is_multisite() && function_exists('get_network_option')) ? get_network_option(null, 'wordfence_version', '0.0.0') : get_option('wordfence_version', '0.0.0'));
+		if (is_multisite() && function_exists('update_network_option')) {
 			update_network_option(null, 'wordfence_version', WORDFENCE_VERSION); //In case we have a fatal error we don't want to keep running install.	
 		}
 		else {
@@ -700,6 +701,20 @@ SQL
 			$wpdb->query("ALTER TABLE {$fileModsTable} ADD COLUMN `isSafeFile` VARCHAR(1) NOT NULL  DEFAULT '?' AFTER `stoppedOnPosition`");
 		}
 		
+		//6.3.7
+		$hooverTable = wfDB::networkPrefix() . 'wfHoover';
+		$hostKeySize = $wpdb->get_var($wpdb->prepare(<<<SQL
+SELECT CHARACTER_MAXIMUM_LENGTH FROM information_schema.COLUMNS
+WHERE TABLE_SCHEMA=DATABASE()
+AND COLUMN_NAME='hostKey'
+AND TABLE_NAME=%s
+SQL
+			, $hooverTable));
+		if ($hostKeySize < 124) {
+			$wpdb->query("ALTER TABLE {$hooverTable} CHANGE `hostKey` `hostKey` VARBINARY(124) NULL DEFAULT NULL");
+		}
+		
+		
 		//Check the How does Wordfence get IPs setting
 		wfUtils::requestDetectProxyCallback();
 
@@ -752,7 +767,7 @@ SQL
 		register_activation_hook(WORDFENCE_FCPATH, 'wordfence::installPlugin');
 		register_deactivation_hook(WORDFENCE_FCPATH, 'wordfence::uninstallPlugin');
 
-		$versionInOptions = (is_multisite() ? get_network_option(null, 'wordfence_version', false) : get_option('wordfence_version', false));
+		$versionInOptions = ((is_multisite() && function_exists('get_network_option')) ? get_network_option(null, 'wordfence_version', false) : get_option('wordfence_version', false));
 		if( (! $versionInOptions) || version_compare(WORDFENCE_VERSION, $versionInOptions, '>')){
 			//Either there is no version in options or the version in options is greater and we need to run the upgrade
 			self::runInstall();
@@ -3038,10 +3053,11 @@ SQL
 		} else {
 			$api = new wfAPI($opts['apiKey'], wfUtils::getWPVersion());
 			try {
-				$keyData = $api->call('ping_api_key', array(), array());
+				$postParams = wfDashboard::updatePOSTParams();
+				$keyData = $api->call('ping_api_key', array(), $postParams);
 				if (isset($keyData['dashboard'])) {
 					wfConfig::set('lastDashboardCheck', time());
-					wfDashboard::processDashboardResponse($keyData['dashboard']);
+					wfDashboard::processDashboardResponse($keyData['dashboard'], @base64_decode($postParams['topBlacklist']));
 				}
 			}
 			catch (Exception $e){
@@ -4255,6 +4271,17 @@ HTACCESS;
 				}
 				return array('ok' => 1, 'data' => $data);
 			}
+		}
+		else if ($grouping == 'blacklist') {
+			$data = $dashboard->blacklist7d['counts'];
+			foreach ($data as &$d) {
+				$d['IP'] = esc_html($d['ip']);
+				$d['localCount'] = esc_html(number_format_i18n($d['local']));
+				$d['networkCount'] = esc_html(number_format_i18n($d['network']));
+				$d['countryFlag'] = esc_attr(wfUtils::getBaseURL() . 'images/flags/' . esc_attr(strtolower($d['countryCode'])) . '.png');
+				$d['countryName'] = esc_html($d['countryName']);
+			}
+			return array('ok' => 1, 'data' => $data);
 		}
 		
 		return array('error' => 'Unknown dashboard data set.');
@@ -6041,9 +6068,13 @@ HTML
 		self::status(10, 'info', "SUM_PAIDONLY:" . $msg);
 	}
 	public static function wfSchemaExists(){
-		$db = new wfDB();
-		global $wpdb; $prefix = $wpdb->base_prefix;
-		$exists = $db->querySingle("show tables like '$prefix"."wfConfig'");
+		global $wpdb;
+		$exists = $wpdb->get_col($wpdb->prepare(<<<SQL
+SELECT TABLE_NAME FROM information_schema.TABLES
+WHERE TABLE_SCHEMA=DATABASE()
+AND TABLE_NAME=%s
+SQL
+			, $wpdb->base_prefix . 'wfConfig'));
 		return $exists ? true : false;
 	}
 	public static function isDebugOn(){
@@ -6906,6 +6937,38 @@ LIMIT %d", $lastSendTime, $limit));
 							
 							if (array_key_exists('data', $jsonData) && array_key_exists('watchedIPList', $jsonData['data'])) {
 								$waf->getStorageEngine()->setConfig('watchedIPs', $jsonData['data']['watchedIPList']);
+							}
+							
+							if (array_key_exists('data', $jsonData) && array_key_exists('blacklisted', $jsonData['data'])) {
+								$blacklisted = base64_decode($jsonData['data']['blacklisted']);
+								$statsToUpdate = array();
+								foreach ($dataToSend as $record) {
+									$blockDay = floor($record[0] / 86400);
+									$blockIP = @wfUtils::inet_pton($record[2]);
+									if ($blockIP !== false) {
+										$position = wfUtils::strpos($blacklisted, $blockIP);
+										if ($position !== false && $position % 16 == 0) {
+											if (!isset($statsToUpdate[$blockIP])) {
+												$statsToUpdate[$blockIP] = array();
+											}
+											$statsToUpdate[$blockIP][$blockDay] = 1;
+										}
+									}
+								}
+								
+								global $wpdb;
+								foreach ($statsToUpdate as $blockIP => $days) {
+									$days = array_keys($days);
+									foreach ($days as $d) {
+										$existingCount = $wpdb->get_var($wpdb->prepare("SELECT blockCount FROM {$wpdb->base_prefix}wfBlockedIPLog WHERE IP = %s AND unixday = %d and blockType = 'waf'", $blockIP, $d));
+										if ($existingCount > 0) {
+											wfActivityReport::logBlockedIP(wfUtils::inet_ntop($blockIP), $d, 'blacklist', $existingCount);
+											$wpdb->query($wpdb->prepare("DELETE FROM {$wpdb->base_prefix}wfBlockedIPLog WHERE IP = %s AND unixday = %d and blockType = 'waf'", $blockIP, $d));
+										}
+									}
+									
+									
+								}
 							}
 						}
 					}
